Move socket handling to LAN component
