+++ firmware/components/um1_lan/include/um1_lan.h
@@ -12,6 +12,7 @@
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
 #include "stddef.h"
+#include <stdbool.h>
 #include "esp_netif.h"
 #include "esp_log.h"
 
@@ -21,5 +22,11 @@
 void handle_client(int client_sock);
 void tcp_lan_server_task(void *pvParameters);
 void udp_lan_server_task(void *pvParameters);
+void init_stream_sockets(void);
+void send_tcp_packet(int uart_port, const uint8_t *data, size_t len);
+void send_udp_packet(int uart_port, const uint8_t *data, size_t len);
+
+extern bool tcp_connected;
+extern bool udp_connected;
 
 #endif // UM1_LAN_H
